pyproject.toml
[project]
name = "chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.7",
    "openai-agents>=0.2.5",
    "python-dotenv>=1.1.1",
]
